# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Northflow Research Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BacktestRunStatus = typeof BacktestRunStatus[keyof typeof BacktestRunStatus];


export const BacktestRunStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface BacktestRun {
  id: number;
  symbol: string;
  strategy: string;
  status: BacktestRunStatus;
  /** @nullable */
  totalTrades?: number | null;
  /** @nullable */
  winRate?: number | null;
  /** @nullable */
  totalPnl?: number | null;
  /** @nullable */
  maxDrawdown?: number | null;
  /** @nullable */
  sharpeRatio?: number | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface BacktestRunInput {
  symbol: string;
  strategy: string;
  startDate?: string;
  endDate?: string;
  notes?: string;
}

export type TradeSide = typeof TradeSide[keyof typeof TradeSide];


export const TradeSide = {
  long: 'long',
  short: 'short',
} as const;

export interface Trade {
  id: number;
  runId: number;
  side: TradeSide;
  entryPrice: number;
  exitPrice: number;
  quantity: number;
  pnl: number;
  /** @nullable */
  fee?: number | null;
  entryTime: string;
  exitTime: string;
}

export interface ResearchSummary {
  totalRuns: number;
  completedRuns: number;
  failedRuns: number;
  avgWinRate: number;
  avgPnl: number;
  bestPnl: number;
  worstPnl: number;
  topSymbols: string[];
}