# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Northflow Research Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: research
    description: Backtest and research run operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /research/runs:
    get:
      operationId: listRuns
      tags: [research]
      summary: List all backtest runs
      responses:
        "200":
          description: List of runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BacktestRun"
    post:
      operationId: createRun
      tags: [research]
      summary: Create a new backtest run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestRunInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestRun"

  /research/runs/{id}:
    get:
      operationId: getRun
      tags: [research]
      summary: Get a specific backtest run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestRun"
        "404":
          description: Not found
    delete:
      operationId: deleteRun
      tags: [research]
      summary: Delete a backtest run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /research/runs/{id}/trades:
    get:
      operationId: listRunTrades
      tags: [research]
      summary: Get trades for a specific run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"

  /research/summary:
    get:
      operationId: getResearchSummary
      tags: [research]
      summary: Aggregated summary across all runs
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    BacktestRun:
      type: object
      required: [id, symbol, strategy, status, createdAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        strategy:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        totalTrades:
          type: ["integer", "null"]
        winRate:
          type: ["number", "null"]
        totalPnl:
          type: ["number", "null"]
        maxDrawdown:
          type: ["number", "null"]
        sharpeRatio:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    BacktestRunInput:
      type: object
      required: [symbol, strategy]
      properties:
        symbol:
          type: string
        strategy:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        notes:
          type: string

    Trade:
      type: object
      required: [id, runId, side, entryPrice, exitPrice, quantity, pnl, entryTime, exitTime]
      properties:
        id:
          type: integer
        runId:
          type: integer
        side:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: number
        quantity:
          type: number
        pnl:
          type: number
        fee:
          type: ["number", "null"]
        entryTime:
          type: string
        exitTime:
          type: string

    ResearchSummary:
      type: object
      required: [totalRuns, completedRuns, failedRuns, avgWinRate, avgPnl, bestPnl, worstPnl, topSymbols]
      properties:
        totalRuns:
          type: integer
        completedRuns:
          type: integer
        failedRuns:
          type: integer
        avgWinRate:
          type: number
        avgPnl:
          type: number
        bestPnl:
          type: number
        worstPnl:
          type: number
        topSymbols:
          type: array
          items:
            type: string