# lib/api-zod/src/generated/types/backtestRun.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Northflow Research Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { BacktestRunStatus } from './backtestRunStatus';

export interface BacktestRun {
  id: number;
  symbol: string;
  strategy: string;
  status: BacktestRunStatus;
  /** @nullable */
  totalTrades?: number | null;
  /** @nullable */
  winRate?: number | null;
  /** @nullable */
  totalPnl?: number | null;
  /** @nullable */
  maxDrawdown?: number | null;
  /** @nullable */
  sharpeRatio?: number | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}