# lib/api-zod/src/generated/types/backtestRunInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Northflow Research Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BacktestRunInput {
  symbol: string;
  strategy: string;
  startDate?: string;
  endDate?: string;
  notes?: string;
}