# lib/api-zod/src/generated/types/researchSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Northflow Research Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchSummary {
  totalRuns: number;
  completedRuns: number;
  failedRuns: number;
  avgWinRate: number;
  avgPnl: number;
  bestPnl: number;
  worstPnl: number;
  topSymbols: string[];
}