# lib/api-zod/src/generated/types/trade.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Northflow Research Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { TradeSide } from './tradeSide';

export interface Trade {
  id: number;
  runId: number;
  side: TradeSide;
  entryPrice: number;
  exitPrice: number;
  quantity: number;
  pnl: number;
  /** @nullable */
  fee?: number | null;
  entryTime: string;
  exitTime: string;
}